Refactored entities, so we can have the new item class. It is a subclass of tile but will have some additional interaction functions.
Added test health potion and shop items.

Changed map.py, so it supports multiple item and enemy generations.

Changed sprite lists, so we have a wall sprite list for the melee shader and a tile sprite list for drawing and everything else (walls are still a part of the tile sprite list).