Move sprite_scale and sprite_size to entity.py.
Created IndicatorBar class and made health and armour bars for the enemy.

There is still a bug though. If an IndicatorBar.full_box.width attribute is set to 0.0 then is set to something between 0.0 and 1.0, a ZeroDivisionError is raised since arcade's width.setter property calculates the new scale but the denominator is set to 0.0 from the previous call. Here's an example: https://discord.com/channels/458662222697070613/766795048049442836/963805904585781298