Added -I to tox extension building and reformatted a couple commands.
Switched to using the extension VectorField in game_view.py. It fully works now except for a small bug which causes a crash. I think it's a problem with the extension's reference counts, and it can probably be solved by switching to just using functions instead of a new type (which was my original aim).

Added dummy object to extensions/__init__.py which will fake the import of the un-compiled extensions and raise a NotImplementedError if code attempts to run it.