Moved hades_extensions to src/ so now C++ code and stub files are separate.
Changed Registry::create_game_object, so it now just creates the game object. This lead to the creation of add_component which adds a component to a given game object. Also changed the add_system method to work similarly to add_component.

Added all bindings for game_objects/ and started working on stub files. Just have a bug with create_game_object.

Modified a few definitions, includes, docs and comments to be more up to date and improved.

Renamed Inventory::capacity() to Inventory::get_capacity().