Changed add_component to add_components to fix the binding.cpp error.
Made SystemBase a class and added get_registry() to access the registry attribute. Also changed the constructor to use a pointer instead of a reference as well as making all update() methods const.

Modified all components to just hold data and not have any methods. If a struct has any methods, it becomes a class.

Simplified the AttackSystem design to not use AttackResult.

Updated a few tests that only worked in Debug mode, but failed in Release mode.

Solved a few clang-tidy and cppcheck linting errors, but there's still more to go.

Modified a few comments to be easier to read and understand.

Added const to a few other variables and optimised various bits.