Updated dependencies.
Started transitioning the Python project to use the new C++ ECS.

Nearly completed the stub files for the C++ project. The bindings do need a bit of rearranging to make them easier to use and to remove unnecessary stuff.

There's just a few registry methods left to bind, but that may require a redesign.