Fixed the binding code, so now it takes in Python type objects and correctly works. This did require removing some attributes for some components in Python, but they probably won't be used.
Updated all the stub files and rearranged the binding modules, so they're easier to understand.

Updated the Python project to use the new C++ ECS.

Renamed the C++ project's test directory to tests to be more in-line with the Python project.

Added CMakePresets.json and improved .gitignore.

Rewrote RegistryError to provide more helpful error messages.

Added some more tests to the C++ project and made sure all the existing ones pass on MSVC and GCC.