Updated dependencies.
Decided against the Python bindings tests and modifying the lint jobs for now. If they become a problem later on, then I can take another look at them.

Removed some constants in constants.py as they were moved to the C++ project.

Removed shop.py as it was empty and will be rewritten at a later date.

Improved the constructor system, but I'm not too sure about the initialisation logic in game.py.