Updated dependencies.
Reverted the change to start_menu.py which caused the buttons to not appear.

Stopped running the Python tests on ubuntu-latest for now due to the EGL bug. This will be fixed at a later date.

Fixed the bug where CodeQL does not scan binding.cpp by manually compiling instead of using autobuild.