Implemented the ranged attack in Python and added functionality to enable player rotation.
Updated the stub files to work with the API changes in the C++ module.

Removed GameObjectType and switched to the C++ version.

The components can't be updated to use the KinematicComponent yet since there is a major bug that crashes the game when initialising multiple walls (since they all share the same wall body in Chipmunk2D). This should be fixed upon rewriting the constructors to use a factory pattern.