Implemented the observer patten in the registry allowing the C++ module to send events to the Python module.
Refactored AttackSystem::do_attack() to use the new observer pattern and send an event BulletCreation to the Python module.

Refactored update() in HadesSprite so it now just updates the position, and the game object deletion is handled in the on_game_object_death() event in the game view. This new observer pattern has introduced some tech debt in the Python module, but this will be refactored later on.

Reformatted a few files to be acceptance with Clang-format.