Added an `InventoryItemButton` to contain all the logic for clicking, changing the sprite, and fetching the sprite data. This could eventually be turned into a factory design for the upgrades layout if I wanted to.
Modified the `PaginatedGridLayout` to have the up and down buttons integrated into it and to correctly update the grid when the up and down buttons are pressed.

Refactored `HadesSprite` to pass the constructor as a parameter instead of the game object type. This simplifies various parts of the sprite and allows for properties to access attributes of the constructor. This has, however, required a BULLET constructor and sprite due to its incompatibility with the constructor design.

Added a back button to the player view and redesigned parts of its API to be more efficient and flexible, meaning it is now correctly working and displaying the correct UI.

Modified constructors to accept a new name and description parameter in the JSON template.

Made a few C++ members read only to Python and made `Inventory::items` accessible for the player view.

Fixed bug where an empty attack_algorithms collection causes a segfault.

Added the `resources` directory as an Arcade resource handle.