Fixed a few typing issues with the player view, so now the upgrades button takes the data during initialisation instead of afterwards. This also involved a few other changes and refactors to improve the overall code quality.
Rewrote the majority of `test_player.py` to work with the new player view architecture. These tests are structured and designed much better than before, making them easier to comprehend.

Made the `window` fixture not use a global variable by switching to a session-scoped fixture that will persist the window across the entire test run. This way the `window` fixture just has to reset the session window every time it is initialised.

`Money::money` is now readwrite instead of readonly.