Introduced the dungeon generation with a random seed each reset. The AI can now train generically to avoid the walls in randomly generated level improving its capabilities.
Moved `wall_distances()` to the `PhysicsEngine` removing the need for getting the space.

Enabled saving of the graphs every `SAVE_INTERVAL`. This combines with the functionality of the video recording to create a new directory for each interval.

Tuned hyperparameters to reduce memory footprint while maintaining performance.

Introduced a new `epsilon` attribute to `DQNAgent` to remove the need for `eps_threshold`.