Implemented the tests for `GameEngine`, `load_hitbox()`, `get_factories()`, and `MapGenerator::get_enemy_limit()`.
Fixed and improved the Python tests so that they work with the new `GameEngine` class. This required moving to a more mock-based design as the Python tests cannot initialise the components any more.

`PhysicsSystem::add_bullet()` now uses the factories to initialise a bullet allowing all of them to be defined in one place.

`load_hitbox()` now returns whether the load was successful or not.