Implemented the enemy avoidance behaviour adding a few more observations and simplifying/modularising the existing reward function. This also required adding a custom event loop due to the ongoing problem with the Pyglet event loop.
Added `GameEngine::get_level_constants()` to allow the grid width and height and enemy limit to be retrieved.

Added a `app_exit()` function to override the main game's functionality of trying to exit the game when the player dies or reaches the goal.

Added a collision handler between the player and wall to delete the player if they phase through the wall and become stuck.