Implemented the player attacking behaviour which required reverting to the Pyglet event loop as the custom event loop broke the player progress bars.
Moved registry updating to `on_fixed_update()` due to problems in the physics system with having a variable delta time.

`AttackSystem::do_attack()` now returns a boolean indicating if the attack was successful or not.

`Game.on_mouse_press()` now returns a boolean indicating if the event was handled or not (more specifically if the attack was successful or not).

Added a `time_since_last_attack` property to `Attack`'s bindings.