Fixed the Pyglet event issue. Turns out it was because the `Game` view was being recreated each episode, leaving dangling references in the event system which accumulated over time making `clock.tick()` take exponentially longer each episode as it tries to call non-existent scheduled functions.
Introduced a `setup()` method to `Game`, `GameUI`, and `PlayerView` to allow them to reset their internal state ready for the next episode. This can also be used to allow multiple plays of a game to happen without recreating the `Game` later on.

Added an option to allow a custom seed to be passed to `setup()` which can be used for the `GameEngine` random generator.