Fixed the intermittent `TestRegistryPlayerEnemyBulletCollision` and `TestRegistryEnemyPlayerBulletCollision` tests as they were initialising the Chipmunk2D shape with a zero-sized vector which is undefined behaviour. This pattern was also repeated in many other tests, so they could have all experienced this intermittent problem.
Switched `DamageSystem::deal_damage()` to instead mark the game objects for deletion as it was thought that these were causing a race condition in Chipmunk2D where game objects may not be immediately deleted after updating leading to the intermittent problem. However, this version was preferred to the previous version, so it's being kept. This does have the side effect of requiring the registry to be updated if you want to delete a game object, but this matches with the real-life workflow of the ECS.

Simplified the Chipmunk2D collision handlers as we no longer require the post-step callbacks due to the new mark for deletion feature.