Redesigned the `Game` view to use the MVC pattern with the main `Game` view now passing functionality off to the `GameController` which interacts with the data stored in the `GameModel` to display the correct functionality in the `GameView`. This new architecture makes it much easier to implement new features and tests in the future.
Added `make_sprite()` to abstract away the sprite class selection process allowing the game to focus on creating the game object.

Removed the `position` parameter from `HadesSprite.__init__()` as this was unnecessary and would have been set during `on_update()` anyway.

Added a `ViewType` enum to reduce the possibility of typos when adding and/or getting views from the window.