Refactored `GameEngine` out into `InputHandler` (to handle input events), `GameState` (to handle game state functionality), and `GameEngine` (to act as the controller between these dependencies).
Implemented a `SaveManager` so the player can save, load, delete, and start new games.

Rewrote `factories.hpp` to instead return an instantiated vector of components through `get_game_object_components()` instead of returning all the factories. This also allows the level initialisation to be encapsulated within the factories and not exposed to the `GameEngine`.

Moved `GameEngine::use_item()` to `InventorySystem::use_item()` and `GameEngine_setup_shop()` to `ShopSystem::add_offerings()` as they are more suited as system functions instead of being in the `GameEngine`.

Added `LoadGameMenuScene` allowing the player to view, load, and delete save files. This also required a few changes to `StartMenu` to allow it to show this new scene.

This feature is nearly complete, but, there is still a bug remaining where the player sprite is not created despite the user being able to control it.